KOGITO-5885  Public API: Refactor Common Service Logic to Shared Classes
as a first approximation, I am creating delegate `Impl` classes for each service;

- quarkus/.../kogito-quarkus-decisions/src/.../DecisionServiceImpl.java
- quarkus/.../kogito-quarkus-predictions/src/.../PredictionServiceImpl.java
- quarkus/.../kogito-quarkus-processes/src/.../StraightThroughProcessServiceImpl.java
- quarkus/.../kogito-quarkus-rules/src/.../RuleUnitServiceImpl.java

these delegates are called respectively by:

- quarkus/.../kogito-quarkus-decisions/src/.../QuarkusDecisionService.java
- quarkus/.../kogito-quarkus-predictions/src/.../QuarkusPredictionService.java
- quarkus/.../kogito-quarkus-processes/src/.../QuarkusStraightThroughProcessService.java
- quarkus/.../kogito-quarkus-rules/src/.../QuarkusRuleUnitService.java

The delegates do not use any Quarkus-related class, so we should be able to safely move them
to a `common` directory when we introduce Spring support.
