+++ quarkus/extensions/kogito-quarkus-processes-extension/kogito-quarkus-processes/src/main/java/org/kie/kogito/core/process/incubation/quarkus/support/QuarkusStraightThroughProcessService.java
@@ -16,46 +16,30 @@
 
 package org.kie.kogito.core.process.incubation.quarkus.support;
 
-import java.util.Map;
+import org.kie.kogito.incubation.common.DataContext;
+import org.kie.kogito.incubation.common.Id;
+import org.kie.kogito.incubation.processes.services.StraightThroughProcessService;
+import org.kie.kogito.process.Processes;
 
+import javax.annotation.PostConstruct;
 import javax.enterprise.context.ApplicationScoped;
 import javax.enterprise.inject.Instance;
 import javax.inject.Inject;
-
-import org.kie.kogito.Model;
-import org.kie.kogito.incubation.common.DataContext;
-import org.kie.kogito.incubation.common.Id;
-import org.kie.kogito.incubation.common.LocalUri;
-import org.kie.kogito.incubation.common.MapDataContext;
-import org.kie.kogito.incubation.common.objectmapper.InternalObjectMapper;
-import org.kie.kogito.incubation.processes.LocalProcessId;
-import org.kie.kogito.incubation.processes.services.StraightThroughProcessService;
-import org.kie.kogito.process.Process;
-import org.kie.kogito.process.ProcessInstance;
-import org.kie.kogito.process.Processes;
 
 @ApplicationScoped
 public class QuarkusStraightThroughProcessService implements StraightThroughProcessService {
     @Inject
     Instance<Processes> processesInstance;
+    StraightThroughProcessServiceImpl delegate;
+
+    @PostConstruct
+    void startup() {
+        this.delegate = new StraightThroughProcessServiceImpl(processesInstance.get());
+    }
 
     @Override
     public DataContext evaluate(Id processId, DataContext inputContext) {
-        Processes processes = processesInstance.get();
-        LocalUri processPath = processId.toLocalId().asLocalUri();
-        if (processPath.startsWith(LocalProcessId.PREFIX)) {
-            LocalProcessId pid = (LocalProcessId) processId;
-            Process<? extends Model> process = processes.processById(pid.processId());
-            MapDataContext mdc = inputContext.as(MapDataContext.class);
-            Class<? extends Model> modelType = process.createModel().getClass();
-            Model model = InternalObjectMapper.objectMapper().convertValue(mdc, modelType);
-            ProcessInstance<? extends Model> instance = process.createInstance(model);
-            instance.start();
-            Map<String, Object> map = instance.variables().toMap();
-            return MapDataContext.of(map);
-        } else {
-            throw new IllegalArgumentException("Not a valid processId " + processPath);
-        }
+        return delegate.evaluate(processId, inputContext);
     }
 
 }
